+++ Member.h
@@ -6,11 +6,11 @@ public:
     Member(unsigned int geneCount, const char* data);
     const char* getData();
     unsigned int getGeneCount();
-    void setFitness(double fitness);
-    double getFitness();
+    void setFitness(int fitness);
+    int getFitness();
 
 private:
     const char* data;
-    double fitness;
+    int fitness;
     unsigned int geneCount;
 };
